+++ src/features/XIT/SET/BFR.vue
@@ -113,7 +113,6 @@ function deleteRule(rule: [string, number, number]) {
     <thead>
       <tr>
         <GripHeaderCell />
-        <th />
         <th>
           <InlineFlex>
             Command